requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr>=1.6
anyjson>=0.3.3
appdirs>=1.3.0 # MIT License
dogpile.cache>=0.5.4
cliff>=1.14.0          # Apache-2.0
httplib2>=0.7.5
lxml>=2.3
oslo.i18n>=1.5.0 # Apache-2.0
oslo.utils>=2.0.0 # Apache-2.0
PrettyTable<0.8,>=0.7
python-keystoneclient>=1.6.0
python-openstackclient>=1.5.0
six>=1.9.0
